refactor: move readUrl and saveUrl utilities to lib/tauri for better organization
