+++ components/instance.tsx
@@ -5,8 +5,8 @@ import { useEffect } from 'react'
 import { Button } from '@/components/ui/button'
 import { Input } from '@/components/ui/input'
 import { Label } from '@/components/ui/label'
+import { readUrl } from '@/lib/tauri/readUrl'
 import { useStore } from '@/store/instance'
-import { readUrl } from '@/utils/readUrl'
 import { startInstance } from '@/utils/startInstance'
 
 export function Instance() {
