+++ utils/startInstance.ts
@@ -1,5 +1,5 @@
+import { saveUrl } from '@/lib/tauri/saveUrl'
 import { normalizeUrl } from '@/utils/normalizeUrl'
-import { saveUrl } from '@/utils/saveUrl'
 
 export const startInstance = async (instanceUrl: string | null) => {
   const normalized = normalizeUrl(instanceUrl)
